+++ packages/ui/src/components/va-collapse/CHANGELOG.md
@@ -4,3 +4,4 @@
 - Added `content` slot
 - Added `header` and `header-content` slot with `iconAttrs`, `attrs` and `value` arguments
 - Improve collapse animation
+- Now stateful by default